requirements.txt
lark-parser==0.10.1
python-hcl2==3.0.5
